[Refactor] 비밀번호 변경화면의 Api호출 메소드 정리 - 불필요한 API호출 제거 / [Chore] 패키지 정리 - 라우터에 등록된 vue파일은 views디렉토리 하위에서 관리하고 그외의 vue파일은 components디렉토리에서 관리
