+++ src/views/authUser/AccountSignOutPage.vue
@@ -1,0 +1,15 @@
+<template>
+  <AccountSignOutComp />
+</template>
+
+<script>
+import AccountSignOutComp from "@/components/authUser/AccountSignOutComp.vue";
+
+export default {
+  components: {
+    AccountSignOutComp,
+  },
+};
+</script>
+
+<style></style>
