+++ src/views/authUser/PasswordUpdatePage.vue
@@ -1,0 +1,14 @@
+<template>
+  <PasswordUpdateComp />
+</template>
+
+<script>
+import PasswordUpdateComp from "@/components/authUser/PasswordUpdateComp.vue";
+export default {
+  components: {
+    PasswordUpdateComp,
+  },
+};
+</script>
+
+<style></style>
