--- src/views/loginUser/SignOutAccount.vue
@@ -1,17 +1,0 @@
-<template>
-  <SignOutAccount/>
-</template>
-
-<script>
-import SignOutAccount from '@/components/loginUser/SignOutAccount.vue';
-
-export default {
-    components:{
-        SignOutAccount
-    }
-}
-</script>
-
-<style>
-
-</style>
