--- src/views/loginUser/UpdatePassword.vue
@@ -1,16 +1,0 @@
-<template>
-  <UpdatePassword/>
-</template>
-
-<script>
-import UpdatePassword from '@/components/loginUser/UpdatePassword.vue';
-export default {
-    components:{
-        UpdatePassword
-    }
-}
-</script>
-
-<style>
-
-</style>
