--- src/views/myTeam/MyTeamRecord.vue
@@ -1,15 +1,0 @@
-<template>
-  <div>
-      
-  </div>
-</template>
-
-<script>
-export default {
-
-}
-</script>
-
-<style>
-
-</style>
